requirements.txt
google-generativeai>=0.3.0
pyperclip>=1.8.0
pathspec>=0.11.0
python-dotenv>=1.0.0
requests>=2.28.0
keyring>=24.0.0
colorama>=0.4.6
